Tarefa1

Description

Intrdução / Objetivos

Face ao grande ojetivo desta tarefa de construir um mapa para o pacman , utilizaram-se estratégias que permitissem gerar labirintos válidos de acordo com o enunciado

Conclusão /Análise

Assim ,após a realização de diversas funções para que o mesmo funcionasse, os maiores desafios foram a implementação dos túneis , bem como a casa dos fantasmas . Estes foram superados com sucesso , e o objetivo desta tarefa foi concluído .

Documentation

generateRandoms :: Int -> Int -> [Int]

gera um dado numero de aleatórios

convertePeca :: Int -> Piece

dado um número converte-o numa peça

printCorridor :: Corridor -> String

converte um corredor por uma string com n no final

converteCorredor :: [Int] -> Corridor

Converte uma lista de inteiros num corredor

converteLabirinto :: [[Int]] -> Maze

Converte uma lista de corredores num labirinto

calcularcelulas :: Int -> Int -> Int

calcular o n de celulas para as quais se tem gerar nº aleatorios

gerarali :: Int -> Int -> Int -> [Int]

gera o total de nº para o labirinto todo exceto as paredes | s = seed introduzida

geraChao :: Int -> Corridor

gera corredor só com peças do tipo (Empty) de tamanho definido

geraParedes :: Int -> Corridor

gera corredor só com peças do tipo (Wall) de tamanho definido (Int)

geraCorredoresMeio :: [Int] -> Corridor

dada uma lista de números aleatórios converte-os num corredor válido , com paredes nos extremos

geranCorredores :: Int -> [Int] -> Maze

n=comprimento interior sem contar com as paredes :total=(n+2) | introduzindo o n , "separa" os elementos da lista x em listas de pecas com n elemntos

chamarcorredores :: Int -> Int -> Int -> Maze

Gera o interior com valores random

calcularmeio :: Maze -> Int

calcula o meio de um labirinto em altura

injetaTuneis :: Int -> Int -> Int -> Maze -> Maze

meio= meio do labirinto ,t= contador (diz a linha para ser alterada) onde serão colocados os túneis, h= altura do maze , lista l= labirinto

substituiEmpty :: Corridor -> Corridor

substitui o 1º e o ultimo elemento de um corredor por peças do tipo (Empty),que será usado ao gerar os túneis

n :: Maze -> Maze

funçao que imprime tuneis nas posiçoes centrais

homeLine1_5 :: Corridor -> Corridor

Supondo que a casa dos fantasmas tem 5 de altura contando com os corredores (Empty) | dado um corredor e substitui parte dele pela 1º5º linha da casa dos fantasmas

homeline2 :: Corridor -> Corridor

dado um corredor substitui parte dele pela 2º linha da casa dos fantasmas

homeline3 :: Corridor -> Corridor

dado um corredor substitui parte dele pela 3º linha da casa dos fantasmas

homeline4 :: Corridor -> Corridor

dado um corredor substitui parte dele pela 4º linha da casa dos fantasmas

procurarInicio :: Int -> Int

procura a posiçao inicial da casa contado com os espaços vazios que a envolvem

substituiForLine1 :: Int -> Int -> Int -> Maze -> Maze

h altura , meio=meio do labirinto em altura ,t =contador

generateMaze :: Int -> Int -> Int -> Maze

Funçao final que chama todas as funçoes necessarias para gerar o labirinto na forma final desejada