Tarefa2

Description

Introdução

Nesta parte do trabalho o objetivo final foi implementar todo o tipo de jogadas do pacman , e posteriormente as dos ghosts para que as funções das tarefas 4 e 5 ffuncionássem de forma correta.

Estratégias

Optou -se por criar diversas funções auxiliares que realizassem as ações no pacman e nos ghosts , como por exemplo

 1.Alterar estados  2.Alterar coordenadas , bem como o labirinto  3.lterar todo o tipo de pontuação no pacman ( pontos por ingerir comidas ,perder vida)

Por fim o grupo chamou todas estas funções na função principal “play” .

Conclusão

A estrutura da tarefa 2 foi alterada após a primeira entrega do trabalho de modo a que as funções desta tarefa pudessem receber jogadas quer para os fantasmas quer para o pacman , de modo a que as funções das tarefas 4,5 e 6 consigam funcionar .Assim a função principal “play” chama duas funções auxiliares “playghost” e “ playpac” , que tal como o nome sugere aplicam jogadas e as suas consequências ,respetivamente , aos fantasmas e ao pacman .

Documentation

findPlayer :: [Player] -> Int -> Player

dada uma lista de jogadores e o id de um jogador devolve o jogador com o respetivo id

indicePlayer :: [Player] -> Int -> Int -> Int

dada uma lista de jogadores um inteiro e um contador "0" dá a posição desse elemento na lista de jogadores

play :: Play -> State -> State

função final alterada , que aplica jogadas tanto a fantasmas como a pacman

testafantasma :: Player -> Bool

Função que nos diz se o a jogada que estamos a aplicar é a um fantsma ou não

playghost :: Play -> State -> State

função que decide se o fantasma pode jogar ou não mediante se existe ou não parede .

playpac :: Play -> State -> State

função final que testa as diversas possibilidades como orientações , novas coordenadas , modo do pacman ,coordenadas coincidentes

getcol :: Player -> Int

dado um jogador dá-nos o inteiro correspondente à coluna que corresponde à coluna do jogador

ghost :: Play -> Coords -> State -> State

função que aciona todas as ações que podem ser relizadas com fantasmas

alteraGhosts :: Coords -> Maze -> [Player] -> [Player]

função que altera as coordenadas do fantasma para a casa caso este tenha sido "comido" pos um pacman

perdevida :: Play -> [Player] -> [Player]

função que faz com que caso o pacman tenha as mesmas coordenadas de um fantasma e esteja no estado normal perca uma vida ou passe para Dying

pontospac2 :: Play -> [Player] -> [Player]

função que dá os pontos ao pac caso coma um ghost e esteja no estado mega

alterastateGhost :: [Player] -> [Player]

função que altera o estado dos fantasmas para dead e a velocidade para metade

checkCoords :: Coords -> [Player] -> Bool

função que verifica se ha algum fantasma na lista de jogadores com as mesmas coordenadas do pacman que ja são dadas

getGhostMode :: Player -> GhostMode

dado um jogador dá-nos o ghostmode ou seja se está Dead ou Alive

ghostToHouse :: Maze -> Coords

função que dado um labirinto nos dá as coordenadas dentro da casa dos fantasmas

novaCord :: Coords -> Play -> Maze -> Coords

função que dadas umas coordenadas e uma jogada nos dá as novas coordenadas

novacoordenadapacman :: Player -> Play -> Maze -> Player

função que nos dá o player com os pontos atualizados,e coordenadas onde a orientação da play e do pacman já foi testada anteriormente

encontra :: (Int, Int) -> Maze -> Piece

função que encontra a peça correspondente a coordenadas dadas no labirinto

testaParede :: Play -> (Int, Int) -> Maze -> Bool

função que testa se é possivel movimentar dadas uma Play e umas coordenadas

coordenadaEtrocaOrientacao :: Player -> Play -> Player

função que troca a orientaçaõ de um jogador caso esta seja diferente da orientação da jogada

pacToEmpty :: Player -> Play -> Maze -> Maze

função que substitui a casa onde o pacman estava por uma casa vazia

posPacEmpty :: Coords -> Maze -> Maze

função que altera o maze e atualiza a posição anterior do fantasma no mapa

cordtunelr :: Player -> Player

função que altera as coordenadas de um player caso esteja em posição de passar o tunel da direita

cordtunell :: Player -> Maze -> Player

função que altera as coordenadas de um player caso esteja em posição de passar o tunel da esquerda

js :: [Player]

jogador1 :: Player

jogador3 :: Player

jogador2 :: Player

maze3 :: [[Piece]]