Tarefa3

Description

Introdução

O objetivo desta tarefa era compactar um labirinto .

Conclusão/Estratégias

Começou-se por compactar horizontalmente e posteriormente verticalmente (através do uso do “repeat “ ,que comparava a linha em que se encontra com as linhas anteriores ). A compactação horizontal foi feita a partir de funções que contavam o número de elementos iguais na mesma linha .

Documentation

groupLinha :: [Piece] -> [[Piece]]

Função que agrupa elementos iguais na mesma linha do maze

contagemGrupinho :: [Piece] -> [(Int, Piece)]

Função que junta as várias peças de uma linha e devolve no estado (nº de vezes que ocorre,piece)

contagemLinha :: [[Piece]] -> [(Int, Piece)]

Função que conta o nº de grupinhos numa linha

contagemMaze :: [[Piece]] -> [[(Int, Piece)]]

Função que aplica a função "contagemLinha" a todas as linhas do maze

comparaLista :: [Piece] -> [Piece] -> Bool

Função que compara se dois corredores são iguais

indicador :: Maze -> Corridor -> Int -> Int

Função que indica o indice da linha do maze ,onde o corredor é igual ao fornecido , primeira linha = indice 0, c = corredor, k = contador (começa em 0)

forLast :: Maze -> Int -> Instructions

Funçao que compara o ultimo elemento da lista com os outros e devolve o repeat (indice da linha) ou a instruct ...

compactMaze :: Maze -> Instructions

Funçao final

maze2 :: [[Piece]]

maze1 :: [[Piece]]