Tarefa4

Description

Introdução

Esta tarefa consistiu na implementação da passagem do tempo no pacman.Esta tarefa foi a mais díficil para o grupo pois apesar de muitas tentativas , não foi conseguido o totalmente objetivo final ,sendo esta a única tarefa inacabada.

Objetivos

O grande objetivo desta tarefa foi tentar implementar as velocidades de forma correta a par da passagem do tempo .

Apesar das diversas tentativas , o grupo não conseguiu encontrar o erro que impedia o funcionamento na perfeição.

Discussão e Conclusão

A grande limitação desta tarefa foi maioritariamente a aplicação recursiva de funções onde dificilmente ocorre recursividade , pois apesar de todas as funções funcionarem na perfeição para o que lhes é pedido , foi impossível concluir na perfeição esta tarefa .

Em suma,apesar da tarefa não estar concluída e da perspetiva do grupo ,apresenta todas as funções necessárias para que a tarefa ficasse totalmente concluída.

Documentation

defaultDelayTime :: Integer

Tempo padrão

passTime :: Int -> State -> State

Vai alterando o estado com o passar do tempo

aplicajogada :: [Play] -> State -> State

Aplica as jogadas dos fantasmas

playagain :: Int -> State -> Player -> State

Função que repete as jogadas dependendo da velocidade , nas respetivas iterações

analisavel :: Int -> Player -> State -> State

Função que analisa a velocidade e os steps,e aplica jogadas

recAnalisavel :: Int -> [Player] -> State -> State

Aplica recursividade para todos os jogadores através da analisavel

analisator :: State -> State

Função que une as auxiliares que alteram a boca e o estado dos jogadores

alterafanta :: [Player] -> PacMode -> [Player]

Se o pacman estiver em modo Mega, altera o estado dos fantasmas

aux20 :: [Player] -> [Player]

Altera a abertura da boca do pacman

aux1 :: [Player] -> [Player]

Altera o estado,velocidade ou o tempo Mega do pacman

alteraalive :: [Player] -> [Player]

Quando o pacman passa para modo Normal , os fantasmas passam de modo Dead para modo Alive e a sua velocidade passa para o dobro

sacavelo :: Player -> Double

Função que dado um jogador devolve a sua velocidade