Tarefa5

Description

Introdução/Objetivos

A tarefa 5 consistia em implementar as movimentações nos fantasmas , mediante uma escolha da melhor estratégia pelo grupo

Estratégias

Assim e mediante o pedido o grupo decidiu implementar as seguintes estratégias.

Caso o fantasma se encontre na casa dos fantasmas , este sai da mesma.

Caso o pacman se encontre em modo mega os fantasmas usam uma estratégia que consiste em manter o movimento até encontrarem uma parede .

Quando encontra a mesma este obtem o movimento nos sentidos do ponteiro do relogio :

se a sua orientação for para a direita estes ficam orientados para baixo ;

se a sua orientação for para baixo este fica orientado para a esquerda ;

se a sua orientação for para a esquerda este fica orientado para cima ;

se a sua orientação for para cima este fica orientado para a direita ;

Caso o pacman esteja em modo Normal , os fantasmas entram em modo de perseguição em que mantêm o seu movimento até encontrarem uma parede. Neste momento os fantasmas avaliam a posição do pacman e escolhem qual a direção a tomar mediante a posição mais próxima do pacman .

Discussão e Conclusão

Apesar da tarefa 5 estar conluída na perfeição , mediante a estratégia escolhida pelo grupo , estas movimentações não conseguiram ser implementadas na sua totalidade face ao facto da tarefa 4 não estar acabada na sua totalidade.

Documentation

ghostPlay :: State -> [Play]

Função que aplica todas as jogadas a todos os fantasmas

aplicaPLaytoGhost :: [Player] -> [Player]

Funçao que devolve uma lista de jogadores apenas com jogadores do tipo fantasma

aux :: Player -> Bool

Verifica se o player é do tipo pacman ou ghost

auxghostPlay :: State -> [Player] -> [Play]

Função que acplica a todos os fantasmas a função saidatoca

saidatoca :: State -> Player -> Play

Função que retira os fantasmas da casa

decidechaseouscat :: State -> Player -> Play

Função que decide ,mediante o estado do pacman, se entra no modo perseguição ou modo de fuga

findpac :: [Player] -> Player

Auxiliar à decidechaseouscat

chaseMode :: State -> Int -> Play

Função que verifica as condições de perseguição do fantasma

alterajogada :: Player -> Player -> Play

Função que altera a orintação do ghost quando este encontra uma parede (no modo de perseguição)

scatterMode :: State -> Int -> Play

Função que gera jogadas para os fantasmas da lista (em caso de fuga)

alterafuga :: Player -> Play

Roda o fantasma numa direção no sentido dos ponteiros do relógio

momentoficadead :: Player -> Play

Função que troca a orientação ao fantasma no momento que o ghost come uma food big