Tarefa6

Description

Introdução / Objetivos

O grande objetivo da tarefa 6 era implementar um robot (bot) que jogasse pacman da maneira mais eficaz possível e similar a um jogador humano. Assim a tarefa foi feita de modo a implementar funções para os 2 modos do pacman jogáveis , (Normal e Mega ).

Estratégia

Primeiro foram feitas as funções necessárias para que o pacman no modo normal, procurasse incrementar os seus pontos através da ingestão de foods (Big ou little).

Em segundo lugar, e em complemento do primeiro ponto, calculou-se a distância “segura” para o pacman puder procurar “foods”. A distância foi definida pelo grupo como 3 unidades, e caso um ghost se aproximasse demasiado o pacman entrava em modo fuga.

Por último definiram-se as funções necessárias para que o pacman, em modo Mega, perseguisse os fantasmas de modo a incrementar os seus pontos e enviá-los de volta á casa de partida.

Discussão e conclusão

Na opinião do grupo esta foi a tarefa mais divertida de todas, pois levou-nos a pensar as estratégias utilizadas por cada um quando joga o jogo, e implementá-las no nosso código de modo a que o nosso robot se tornasse um bot que conseguisse fazer boas pontuações.

Concluindo o bot podia ter inúmeras estratégias diferentes, pensadas por outros jogadores, mas o nosso executa apenas as estratégias pensadas por cada elemento do grupo.

Documentation

botMovesPac :: Int -> State -> Maybe Play

Executa as varias plays ao jogador do tipo pacman dependo dos estados

findPoints :: Int -> State -> Play

Procura comida (little or big) no mapa de modo a incrementar os seus pontos e passar o nivel se a dist>3 ou foge do ghost

distPacGhost :: Player -> Player -> Float

Calcula a distancia entre o pacman e o ghost mais proximo

runNow :: Player -> Player -> Maze -> Play

Funçao que direciona o pacman de modo a fugir do ghost mais proximo

playcerta :: Player -> Maze -> Play

Desloca-se para a food mais proxima conforme a orientaçao do pacman ( funciona apenas se a comida estiver nas 3 casas mais proximas , U D R L)

procuraNextFood :: Player -> Int -> Maze -> Play

C = contador (começa em 1) e procura a comida mais proxima

testaFoodL :: (Int, Int) -> Maze -> Bool

Testa à esquerda da posiçao do pacman se encontra comida (big or little)

testaFoodR :: (Int, Int) -> Maze -> Bool

Testa à direita da posiçao do pacman se encontra comida (big or little)

testaFoodU :: (Int, Int) -> Maze -> Bool

Testa se em cima da posiçao do pacman se encontra comida (big or little)

testaFoodD :: (Int, Int) -> Maze -> Bool

Testa se em baixo da posiçao do pacman se encontra comida (big or little)

trytoChase :: Int -> State -> Play

Persegue os fantasmas no mapa de modo a incrementar os seus pontos

findGhostclose :: Player -> [Player] -> Player

Devolve o ghost que se encontra mais proximo do pacman no mapa

getGhosts :: [Player] -> [Player]

Recebe uma lista de players e devolve uma lista de jogadores do tipo Ghost

closer :: Coords -> [Player] -> Player

Compara entre os fantasmas qual deles esta mais proximo do pacman

alterajogadas :: Player -> Player -> Play

Direciona o pacman em direçao ao ghost

bot :: Int -> State -> Maybe Play

Funçao final